requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
